CLN: move _version.py back for versioneer
